create ComradeAHMC
